requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.1
attrs==20.3.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-archive==0.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-model-utils==4.1.1
django-stdimage==5.3.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
kombu==5.0.2
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.1.0
pytest-env==0.6.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
